add bulk DataChangeNotificationCallback to client_subscriptions
The goal is to bypass slow list lookup when monitoring thousands of
items. A single callback is introduced, at the subscription level.
When this callback is not NULL, processing of the
dataChangeNotification is delegated to the user-supplied callback,
instead of iterating the list and and calling each monitored item's
callback.